+++ react/country/JAM.ts
@@ -1,0 +1,145 @@
+import { TWO_LEVELS } from '../constants'
+import { secondLevelPostalCodes } from '../transforms/postalCodes'
+import { getOneLevel, getTwoLevels } from '../transforms/addressFieldsOptions'
+import type { PostalCodeRules } from '../types/rules'
+import countryData from './data/JAM.json'
+
+const rules: PostalCodeRules = {
+  country: 'JAM',
+  abbr: 'JM',
+  postalCodeFrom: TWO_LEVELS,
+  postalCodeLevels: ['state', 'city'],
+  secondLevelPostalCodes: secondLevelPostalCodes(countryData),
+  fields: [
+    {
+      hidden: true,
+      name: 'country',
+      maxLength: 100,
+      label: 'country',
+      size: 'medium',
+    },
+    {
+      autoComplete: 'nope',
+      hidden: true,
+      label: 'postalCode',
+      maxLength: 50,
+      name: 'postalCode',
+      postalCodeAPI: false,
+      required: true,
+      regex: /^[Jj][Mm][A-Za-z]{3}\d{2}$/,
+      size: 'small',
+    },
+    {
+      name: 'state',
+      maxLength: 100,
+      label: 'department',
+      required: true,
+      size: 'xlarge',
+      level: 1,
+      options: getOneLevel(countryData),
+    },
+    {
+      name: 'city',
+      maxLength: 100,
+      label: 'municipality',
+      required: true,
+      size: 'xlarge',
+      level: 2,
+      basedOn: 'state',
+      optionsMap: getTwoLevels(countryData),
+    },
+    {
+      hidden: true,
+      name: 'neighborhood',
+      maxLength: 100,
+      label: 'neighborhood',
+      size: 'large',
+    },
+    {
+      name: 'street',
+      label: 'street',
+      required: true,
+      size: 'xlarge',
+    },
+    {
+      hidden: true,
+      name: 'number',
+      maxLength: 750,
+      label: 'number',
+      size: 'small',
+      autoComplete: 'nope',
+    },
+    {
+      hidden: true,
+      name: 'complement',
+      maxLength: 750,
+      label: 'complement',
+      size: 'large',
+    },
+    {
+      hidden: true,
+      name: 'reference',
+      maxLength: 750,
+      label: 'reference',
+      size: 'xlarge',
+    },
+    {
+      name: 'receiverName',
+      elementName: 'receiver',
+      maxLength: 750,
+      label: 'receiverName',
+      size: 'xlarge',
+      required: true,
+    },
+  ],
+  geolocation: {
+    postalCode: {
+      valueIn: 'long_name',
+      types: ['postal_code'],
+      required: true,
+    },
+
+    number: {
+      valueIn: 'long_name',
+      types: ['street_number'],
+      required: false,
+      notApplicable: true,
+    },
+
+    street: { valueIn: 'long_name', types: ['route'] },
+
+    neighborhood: {
+      valueIn: 'long_name',
+      types: ['neighborhood'],
+    },
+
+    state: {
+      valueIn: 'long_name',
+      types: ['administrative_area_level_1'],
+    },
+
+    city: {
+      valueIn: 'long_name',
+      types: ['administrative_area_level_2', 'locality'],
+    },
+
+    receiverName: {
+      required: true,
+    },
+  },
+  summary: [
+    [
+      { name: 'street' },
+      { delimiter: ' ', name: 'number' },
+      { delimiter: ', ', name: 'complement' },
+    ],
+    [
+      { name: 'neighborhood', delimiterAfter: ' - ' },
+      { name: 'city' },
+      { delimiter: ' - ', name: 'state' },
+    ],
+    [{ name: 'postalCode' }],
+  ],
+}
+
+export default rules