+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Jamaica country rules.
+
 ## [3.26.8] - 2022-08-10
 ### Fixed
 - Invalidate whitespace-only required fields.